Update fuzzy: fixed spelling, removed unnecessary casting, FuzzyGameWorld.addPoints parameter was ignored.
